Added Hit Detection Plus Failures
Added two extra timers, one that triggers after 5 seconds to enable the button presses to be tracked. Second timer gives the player 8 milliseconds to click a button, if green they get a point if red they get a life taken away. If they run out of lives the application terminates